CLP: Move Navigation class to its own file (navigation.py)
* Also move the classes that Navigation class exclusively uses
  (namely, QueueCustom and UpdateNavigationScene) to the same file.
